Implemented condense-type
Getting really tired of Scheme though. It is not behaving well in Emacs.
Emacs keeps freezing. I keep going back and forth between trying Guile
and Chicken. Guile seems to have a bit more polish but it's actually
Chicken that works slightly better for me in Emacs. Overall though the
development experience in Emacs is vastly inferior to that of Racket.