delete: drop support for special form [if ...]
As it stands, it doesn't agree super well with the combinator calculus.
Intending to add if back as a function. There are only three possible reasons to
need if as a special form:

 1. Because your language is statement-oriented
 2. Because your language is strict and has ambient side-effects
 3. Because your language is strict and the branch not selected could be
    expensive to construct only to throw away

1) doesn't apply here; 2) by and large Metacore doesn't have ambient side
effects (we haven't strictly eliminated them, but the next point will seal the
deal); 3) you can just thunk your "then" and "else" branches with the K
combinator to totally eliminate points 2 and 3.